# task-manager/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Manager API",
    "description": "API for managing tasks and task-related entities",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API"
    },
    {
      "url": "/api/v1",
      "description": "Task Management API v1"
    }
  ],
...
  "paths": {
...
    "/v1/stories": {
      "get": {
        "tags": ["User Stories"],
        "summary": "Get all user stories",
        "parameters": [
          {
            "name": "epicId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter stories by epic ID"
          },
          {
            "name": "jtbdId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter stories by job to be done ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user stories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserStory"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User Stories"],
        "summary": "Create a new user story",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User story created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStory"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stories/{storyId}": {
      "get": {
        "tags": ["User Stories"],
        "summary": "Get user story by ID",
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User story details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStory"
                }
              }
            }
          },
          "404": {
            "description": "User story not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User Stories"],
        "summary": "Update user story",
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User story updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStory"
                }
              }
            }
          },
          "404": {
            "description": "User story not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["User Stories"],
        "summary": "Delete user story",
        "parameters": [
          {
            "name": "storyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User story deleted successfully"
          },
          "404": {
            "description": "User story not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "UserStory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "epicId": {
            "type": "string"
          },
          "jtbdId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "benefit": {
            "type": "string"
          },
          "acceptanceCriteria": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "points": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["backlog", "ready", "in_progress", "review", "done"]
          },
          "technicalConsiderations": {
            "type": "string"
          },
          "designConsiderations": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserStoryInput": {
        "type": "object",
        "required": ["userType", "action", "benefit"],
        "properties": {
          "epicId": {
            "type": "string"
          },
          "jtbdId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "benefit": {
            "type": "string"
          },
          "acceptanceCriteria": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "points": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["backlog", "ready", "in_progress", "review", "done"],
            "default": "backlog"
          },
          "technicalConsiderations": {
            "type": "string"
          },
          "designConsiderations": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          }
        }
      },
...
    }
  }
} 